moved coveo calls to commerce JS
