+++ blocks/product-overview/product-overview.js
@@ -1,4 +1,4 @@
-import { getProductResponse } from '../../scripts/scripts.js';
+import { getProductResponse } from '../../scripts/commerce.js';
 
 export default async function decorate(block) {
   const response = await getProductResponse();
