+++ blocks/product-parts/product-parts.js
@@ -1,7 +1,7 @@
 import {
   div, img, p,
 } from '../../scripts/dom-builder.js';
-import { getProductResponse } from '../../scripts/scripts.js';
+import { getProductResponse } from '../../scripts/commerce.js';
 
 export default async function decorate(block) {
   const response = await getProductResponse();
